Refactoring: pdfanno.js
