+++ src/pdfanno.js
@@ -6,7 +6,6 @@ import * as pdfDropdown from './page/ui/pdfDropdown';
 import * as primaryAnnoDropdown from './page/ui/primaryAnnoDropdown';
 import * as annoListDropdown from './page/ui/annoListDropdown';
 import * as referenceAnnoDropdown from './page/ui/referenceAnnoDropdown';
-
 import * as annotationsTools from './page/ui/annotationTools';
 
 import {
@@ -16,38 +15,16 @@ import {
     setupColorPicker
 } from './page/util/display';
 
+import {
+    listenWindowLeaveEvent,
+    unlistenWindowLeaveEvent,
+    resizeHandler
+} from './page/util/window';
 
 /**
  * The data which is loaded via `Browse` button.
  */
 window.fileMap = {};
-
-/**
- * Resize the height of elements adjusting to the window.
- */
-function resizeHandler() {
-
-    // PDFViewer.
-    let height = $(window).innerHeight() - $('#viewer').offset().top;
-    $('#viewer iframe').css('height', `${height}px`);
-
-    // Dropdown for PDF.
-    let height1 = $(window).innerHeight() - ($('#dropdownPdf ul').offset().top || 120);
-    $('#dropdownPdf ul').css('max-height', `${height1 - 20}px`);
-
-    // Dropdown for Primary Annos.
-    let height2 = $(window).innerHeight() - ($('#dropdownAnnoPrimary ul').offset().top || 120);
-    $('#dropdownAnnoPrimary ul').css('max-height', `${height2 - 20}px`);
-
-    // Dropdown for Anno list.
-    let height3 = $(window).innerHeight() - ($('#dropdownAnnoList ul').offset().top || 120);
-    $('#dropdownAnnoList ul').css('max-height', `${height3 - 20}px`);
-
-    // Dropdown for Reference Annos.
-    let height4 = $(window).innerHeight() - ($('#dropdownAnnoReference ul').offset().top || 120);
-    $('#dropdownAnnoReference ul').css('max-height', `${height4 - 20}px`);
-
-}
 
 /**
     Adjust the height of viewer according to window height.
@@ -56,79 +33,6 @@ function adjustViewerSize() {
     window.removeEventListener('resize', resizeHandler);
     window.addEventListener('resize', resizeHandler);
     resizeHandler();
-}
-
-function _getDownloadFileName() {
-
-    // The name of Primary Annotation.
-    let primaryAnnotationName;
-    $('#dropdownAnnoPrimary a').each((index, element) => {
-        let $elm = $(element);
-        if ($elm.find('.fa-check').hasClass('no-visible') === false) {
-            primaryAnnotationName = $elm.find('.js-annoname').text();
-        }
-    });
-    if (primaryAnnotationName) {
-        return primaryAnnotationName;
-    }
-
-    // The name of PDF.
-    let pdfFileName = iframeWindow.getFileName(iframeWindow.PDFView.url);
-    return pdfFileName.split('.')[0] + '.anno';
-}
-
-/**
- * Export the primary annotation data for download.
- */
-function downloadAnnotation() {
-
-    window.iframeWindow.PDFAnnoCore.getStoreAdapter().exportData().then(annotations => {
-        let blob = new Blob([annotations]);
-        let blobURL = window.URL.createObjectURL(blob);
-        let a = document.createElement('a');
-        document.body.appendChild(a); // for firefox working correctly.
-        a.download = _getDownloadFileName();
-        a.href = blobURL;
-        a.click();
-        a.parentNode.removeChild(a);
-    });
-
-    unlistenWindowLeaveEvent();
-}
-
-/**
- * Delete all annotations.
- */
-function deleteAllAnnotations() {
-
-    // Comfirm to user.
-    let userAnswer = window.confirm('Are you sure to clear the current annotations?');
-    if (!userAnswer) {
-        return;
-    }
-
-    iframeWindow.annotationContainer.destroy();
-
-    let documentId = window.iframeWindow.getFileName(window.iframeWindow.PDFView.url);
-    window.iframeWindow.PDFAnnoCore.getStoreAdapter().deleteAnnotations(documentId).then(() => {
-        reloadPDFViewer();
-    });
-}
-
-/**
- * Set the confirm dialog at leaving the page.
- */
-function listenWindowLeaveEvent() {
-    $(window).off('beforeunload').on('beforeunload', () => {
-        return 'You don\'t save the annotations yet.\nAre you sure to leave ?';
-    });
-}
-
-/**
- * Unset the confirm dialog at leaving the page.
- */
-function unlistenWindowLeaveEvent() {
-    $(window).off('beforeunload');
 }
 
 /**
@@ -144,21 +48,11 @@ function startApplication() {
         // Adjust the height of viewer.
         adjustViewerSize();
 
-        // Initialize tool buttons' behavior.
-        // initializeAnnoToolButtons();
-        annotationsTools.setup1();
-
         // Reset the confirm dialog at leaving page.
         unlistenWindowLeaveEvent();
     });
 
-    // Set viewMode behavior after annotations rendered.
-    iframeWindow.addEventListener('annotationrendered', () => {
-        // window.iframeWindow.PDFAnnoCore.UI.disableViewMode();
-        window.iframeWindow.PDFAnnoCore.UI.enableViewMode();
-    });
-
-    // Set the confirm dialog at page leaving.
+    // Set the confirm dialog when leaving a page.
     iframeWindow.addEventListener('annotationUpdated', listenWindowLeaveEvent);
 }
 
@@ -175,16 +69,13 @@ window.addEventListener('DOMContentLoaded', e => {
     // Start application.
     startApplication();
 
-    // Setup loading tools for PDFs and Anno files.
-    // setupBrowseButton();
+    // Setup UI parts.
     browseButton.setup();
-    // setupPdfDropdown();
     pdfDropdown.setup();
-
     primaryAnnoDropdown.setup();
     referenceAnnoDropdown.setup();
     annoListDropdown.setup();
-    // setupAnnoListDropdown();
+    annotationsTools.setup();
 
     window.addEventListener('restartApp', startApplication);
 });
